+++ src/main/java/zaftnotameni/sharedbehaviors/ManaMachineComponentBlock.java
@@ -1,0 +1,8 @@
+package zaftnotameni.sharedbehaviors;
+import net.minecraft.world.level.block.Block;
+public class ManaMachineComponentBlock extends Block {
+  public ManaMachineComponentBlock(Properties pProperties) {
+    super(pProperties);
+  }
+
+}